feat(hint): Interaction entity managed. All service divided into Service Interface and ServiceImpl Class. All MongoDB references removed.
