+++ T1-G11/applicazione/manvsclass/docker-compose.yml
@@ -9,8 +9,6 @@ services:
     expose:
       - 8081
     environment:
-      MONGODB_ROOT_PASSWORD: ${MONGODB_ROOT_PASSWORD}
-      MONGODB_ROOT_USERNAME: ${MONGODB_ROOT_USERNAME}
       MAIL_PASSWORD: ${MAIL_PASSWORD}
       OTEL_EXPORTER_OTLP_ENDPOINT: http://otel-collector:4317
       OTEL_EXPORTER_OTLP_PROTOCOL: grpc
@@ -21,7 +19,6 @@ services:
       SPRING_DATASOURCE_PASSWORD: ${POSTGRES_PASSWORD}
     depends_on:
       - postgres-db
-      - mongo_db
     volumes:
       - ./FilesUpload:/Files-Upload
       - VolumeT0:/VolumeT0
@@ -32,20 +29,6 @@ services:
       options:
         max-size: "10m"
         max-file: "3"
-
-  mongo_db:
-    image: "mongo:6.0.6"
-    container_name: t1-mongo_db
-    restart: always
-    ports:
-      - "27017:27017"
-    expose:
-      - "27017:27017"
-    volumes:
-      - ./mongo-init.js:/docker-entrypoint-initdb.d/mongo-init.js:ro
-      - mongodb_data_volume:/data/db
-    networks:
-      - global-network
 
   postgres-db:
     image: "postgres:14-alpine3.17"
