+++ T1-G11/applicazione/manvsclass/src/main/java/com/groom/manvsclass/mapper/HintMapper.java
@@ -4,8 +4,8 @@ import com.groom.manvsclass.model.entity.ClassUTEntity;
 import com.groom.manvsclass.model.entity.HintEntity;
 import com.groom.manvsclass.model.dto.Hint;
 import com.groom.manvsclass.model.dto.HintResponse;
-import com.groom.manvsclass.model.repository.jpa.AdminRepository;
-import com.groom.manvsclass.model.repository.jpa.ClassUTRepository;
+import com.groom.manvsclass.model.repository.AdminRepository;
+import com.groom.manvsclass.model.repository.ClassUTRepository;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Mappings;
