+++ T1-G11/applicazione/manvsclass/src/main/java/com/groom/manvsclass/model/dto/TeamUpdate.java
@@ -1,0 +1,13 @@
+package com.groom.manvsclass.model.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class TeamUpdate {
+    private String idTeam;
+    private String newName;
+}
