+++ T1-G11/applicazione/manvsclass/src/main/java/com/groom/manvsclass/model/entity/TeamEntity.java
@@ -1,12 +1,10 @@
 package com.groom.manvsclass.model.entity;
 
-import com.vladmihalcea.hibernate.type.json.JsonBinaryType;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import org.hibernate.annotations.Type;
-import org.hibernate.annotations.TypeDef;
 
 import javax.persistence.*;
 import java.util.ArrayList;
