+++ T1-G11/applicazione/manvsclass/src/main/java/com/groom/manvsclass/model/repository/OperationRepository.java
@@ -1,4 +1,4 @@
-package com.groom.manvsclass.model.repository.jpa;
+package com.groom.manvsclass.model.repository;
 
 import com.groom.manvsclass.model.entity.OperationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
