--- T1-G11/applicazione/manvsclass/src/main/java/com/groom/manvsclass/model/repository/mongo/InteractionRepositoryMongoDB.java
@@ -1,8 +1,0 @@
-package com.groom.manvsclass.model.repository.mongo;
-
-import com.groom.manvsclass.model.interactionMongoDB;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface InteractionRepositoryMongoDB extends MongoRepository<interactionMongoDB, String> {
-
-}
