--- T1-G11/applicazione/manvsclass/src/main/java/com/groom/manvsclass/model/repository/mongo/MongoRepository.java
@@ -1,5 +1,0 @@
-package com.groom.manvsclass.model.repository.mongo;
-
-public class MongoRepository<T1, T2> {
-
-}
