+++ T1-G11/applicazione/manvsclass/src/main/java/com/groom/manvsclass/service/ClassUTService.java
@@ -1,0 +1,10 @@
+package com.groom.manvsclass.service;
+
+import com.groom.manvsclass.model.entity.ClassUTEntity;
+
+public interface ClassUTService {
+
+    public Long updateClassUT(String oldName, ClassUTEntity newContent);
+
+    public ClassUTEntity deleteClassUT(String name);
+}
