+++ T1-G11/applicazione/manvsclass/src/main/java/com/groom/manvsclass/service/implementation/HintServiceImpl.java
@@ -4,14 +4,13 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.groom.manvsclass.mapper.HintMapper;
 import com.groom.manvsclass.model.entity.AdminEntity;
-import com.groom.manvsclass.model.entity.ClassUTEntity;
 import com.groom.manvsclass.model.entity.HintEntity;
 import com.groom.manvsclass.model.dto.Hint;
 import com.groom.manvsclass.model.dto.HintResponse;
 import com.groom.manvsclass.model.enums.HintTypeEnum;
-import com.groom.manvsclass.model.repository.jpa.AdminRepository;
-import com.groom.manvsclass.model.repository.jpa.HintRepository;
-import com.groom.manvsclass.model.repository.jpa.ClassUTRepository;
+import com.groom.manvsclass.model.repository.AdminRepository;
+import com.groom.manvsclass.model.repository.HintRepository;
+import com.groom.manvsclass.model.repository.ClassUTRepository;
 import com.groom.manvsclass.service.HintService;
 import com.groom.manvsclass.service.JwtService;
 import com.groom.manvsclass.util.HintSpecifications;
