+++ T1-G11/applicazione/manvsclass/src/main/resources/application-dev.properties
@@ -3,5 +3,3 @@ API_GATEWAY_ENDPOINT=127.0.0.1
 
 # Porte servizi
 API_GATEWAY_PORT=8090
-
-spring.data.mongodb.uri=mongodb://localhost:27017/manvsclass
