+++ T1-G11/applicazione/manvsclass/src/main/resources/application.properties
@@ -1,9 +1,5 @@
-spring.data.mongodb.database=manvsclass
 server.port=8081
 spring.messages.basename=lang/messages
-
-spring.data.mongodb.host=localhost
-spring.data.mongodb.port=27017
 
 spring.datasource.url=jdbc:postgresql://localhost:5433/manvsclass
 #spring.datasource.url=jdbc:postgresql://localhost:5433/temp_db_for_migrations
